requirements.txt
fastapi
uvicorn
pandas
openpyxl
groq
python-dotenv
